# core/api/docs/v2/openapi.yaml
components:
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Invalid path, query or request input
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
      description: Missing, unknown, revoked or expired key; shared authentication envelope
...
  schemas:
...
    CreateChatRequest:
      properties:
        name:
          type: string
      type: object
...
  securitySchemes:
    bearerauth:
      scheme: bearer
      type: http
...
info:
  contact:
    email: support@anytype.io
    name: Anytype Support
    url: https://anytype.io/contact
  description: |-
    The agent-oriented Anytype local API. An object is one AnyBlock JSON document rather than a tree of blocks, so a single GET returns a whole editable document and a single PATCH edits it.
    Everything this API names for itself is snake_case: path and query parameters, request and response fields, and the names of the PATCH ops. Things are addressed by name rather than by id: property keys, option names, and `type` as a type key. Object ids remain ids.
    Property and type keys have two spellings, and a served body picks one. The default is the key: a snake_case slug minted when the property was created and frozen from then on, so it survives every rename. Store that spelling in a config, a script, an integration. `?keys=name` on a read serves the display-name vocabulary instead, "Due date" rather than due_date, which is the spelling to show a user their own words. Input needs no such choice: every channel accepts both, plus the forgiving fold between them, so dueDate, due-date, "Due date" and Дата выполнения all address the property they name. A spelling matching two properties is a 400 listing both, and an unknown one never creates anything.
...
    A read never fails on content it cannot represent. Whatever a representation cannot express is reported in `warnings` beside the result.
...
  license:
    name: Any Source Available License 1.0
...
openapi: 3.1.0
paths:
  /v2/auth/whoami:
    get:
      description: This describes the key, not a person; there is one account behind
        this API. Branch on `grant.scoped`. False is a legacy key with no space restriction,
        and its `spaces` list is empty rather than absent. True means the key reaches
        exactly the spaces listed, with the permission listed beside each one.
      operationId: auth_whoami
      parameters:
      - description: 'How grant.spaces[].id is spelled: compact (default) is the short
          space reference; full is the whole <cid>.<replicationKey> id, and the spelling
          to store outside this API'
        in: query
        name: ids
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhoamiResponse'
          description: The key's grant, as it is enforced
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
      security:
      - bearerauth: []
      summary: Describe the calling key
      tags:
      - Auth
  /v2/schemas:
    get:
      operationId: list_schemas
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaIndex'
          description: Schema index
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
      security:
      - bearerauth: []
      summary: List the available schemas
      tags:
      - Schemas
  /v2/schemas/{kind}:
    get:
      operationId: get_schema
      parameters:
      - description: Schema kind, as listed by GET /v2/schemas
        in: path
        name: kind
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaEntry'
          description: Schema + example
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unknown kind
      security:
      - bearerauth: []
      summary: Get the schema for one kind
      tags:
      - Schemas
  /v2/schemas/ops/{op}:
    get:
      description: The example is a single op object, ready to drop into an edit request's
        `ops` array, not a whole request body.
      operationId: get_op_schema
      parameters:
      - description: 'Op name: set_properties, update_block, replace_subtree, insert_blocks,
          move_block, delete_block, replace_text, set_cell, update_view, insert_view,
          move_view, delete_view, add_items, remove_items'
        in: path
        name: op
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaEntry'
          description: Schema + example
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unknown op
      security:
      - bearerauth: []
      summary: Get the schema for one edit op
      tags:
      - Schemas
  /v2/search:
    post:
      description: Type keys and option names are resolved per space. A name that
        resolves in only some spaces searches those and warns about the rest. `total`
        is the sum of the per-space counts, and each row carries its `space_id`.
      operationId: search_global
      parameters:
      - description: Items to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Items to return
        in: query
        name: limit
        schema:
          default: 25
          type: integer
      - description: 'How each row''s space_id is spelled: compact (default) is the
          short space reference; full is the whole <cid>.<replicationKey> id, and
          the spelling to store outside this API'
        in: query
        name: ids
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequestDoc'
        description: Search request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse-ObjectRow'
          description: Minimal object rows with space_id
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
      security:
      - bearerauth: []
      summary: Search every space
      tags:
      - Search
  /v2/spaces:
    get:
      description: Only live spaces are listed. A space that is deleted, left, or
        still joining does not appear.
      operationId: list_spaces
      parameters:
      - description: compact (default) is the short space reference; full is the whole
          <cid>.<replicationKey> id, and the spelling to store outside this API
        in: query
        name: ids
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse-SpaceRow'
          description: Minimal space rows
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
      security:
      - bearerauth: []
      summary: List the account's spaces
      tags:
      - Spaces
    post:
      description: A retry the server already handled makes a second space unless
        it carries the same idempotency key. A dry run validates the body and stops
        there; creating a space cannot be simulated.
      operationId: create_space
      parameters:
      - description: Validate the body without creating
        in: query
        name: dry_run
        schema:
          type: boolean
      - description: compact (default) is the short space reference; full is the whole
          <cid>.<replicationKey> id of the new space, and the spelling to store outside
          this API
        in: query
        name: ids
        schema:
          type: string
      - description: 'Replay guard: the same key with the same body replays the stored
          response'
        in: header
        name: Idempotency-Key
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpaceRequest'
        description: The space to create
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
          description: Dry run; validation result without committing
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
          description: Created space
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "413":
          $ref: '#/components/responses/RequestTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Create a space
      tags:
      - Spaces
  /v2/spaces/{space_id}:
    get:
      description: Only live spaces are served. A space that is deleted, left, or
        still joining is a 404.
      operationId: get_space
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: compact (default) is the short space reference; full is the whole
          <cid>.<replicationKey> id, and the spelling to store outside this API
        in: query
        name: ids
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
          description: The space row
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Space not found
      security:
      - bearerauth: []
      summary: Get one space
      tags:
      - Spaces
    patch:
      description: At least one of the two fields must be present; a field left out
        keeps its current value.
      operationId: update_space
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      - description: compact (default) is the short space reference; full is the whole
          <cid>.<replicationKey> id, and the spelling to store outside this API
        in: query
        name: ids
        schema:
          type: string
      - description: 'Replay guard: the same key with the same body replays the stored
          response'
        in: header
        name: Idempotency-Key
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpaceRequest'
        description: The fields to change
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
          description: The updated space row
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Space not found or not live
        "409":
          $ref: '#/components/responses/Conflict'
        "413":
          $ref: '#/components/responses/RequestTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Update a space
      tags:
      - Spaces
  /v2/spaces/{space_id}/chats:
    get:
      description: A row carries no unread counters. Per-chat unread state comes back
        with the messages read instead.
      operationId: list_chats
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Rows to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Rows to return
        in: query
        name: limit
        schema:
          default: 25
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse-ChatRow'
          description: Chat rows
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Space not found
      security:
      - bearerauth: []
      summary: List the chats in a space
      tags:
      - Chat
    post:
      description: Messages are not blocks. Add them through the messages route; a
        document edit cannot reach them.
      operationId: create_chat
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      - description: 'Replay guard: the same key with the same body replays the stored
          response'
        in: header
        name: Idempotency-Key
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
        description: The chat to create
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResult'
          description: Dry run; validation result without committing
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResult'
          description: Created chat row
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "413":
          $ref: '#/components/responses/RequestTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Create a chat
      tags:
      - Chat
...
  /v2/spaces/{space_id}/chats/{chat_id}/messages/{message_id}/reactions:
    post:
      description: '`added` says which way the toggle went. A dry run predicts it,
        but when there is no account identity to predict with it omits the field and
        says so in `warnings`.'
      operationId: toggle_chat_reaction
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Chat object id
        in: path
        name: chat_id
        required: true
        schema:
          type: string
      - description: Message id
        in: path
        name: message_id
        required: true
        schema:
          type: string
      - description: Report the would-be outcome without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      - description: 'Replay guard: the same key with the same body replays the stored
          response'
        in: header
        name: Idempotency-Key
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatReactionRequest'
        description: The emoji to toggle
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatReactionResult'
          description: Toggle outcome
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Chat or message not found
        "409":
          $ref: '#/components/responses/Conflict'
        "413":
          $ref: '#/components/responses/RequestTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Toggle a reaction on a chat message
      tags:
      - Chat
...
  /v2/spaces/{space_id}/collections/{collection_id}/views:
    get:
      operationId: get_collection_views
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Collection object id
        in: path
        name: collection_id
        required: true
        schema:
          type: string
      - description: Items to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Items to return
        in: query
        name: limit
        schema:
          default: 25
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse-ViewObject'
          description: The stored views, with their sorts, filters and columns
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Wrong-layout target
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Space or collection not found
      security:
      - bearerauth: []
      summary: List a collection's views
      tags:
      - Lists
...
servers:
...